Update additional-sets-merge.py
The mean of all prices over the years is taken, and then it is calculated whether an area is in the 33th or 66th percentile, which means that it can be classified as low, normal or high.
Only thing left to do is connect the areas to the main dataframe area.